strukturerat upp kod
städat och strukturerat om koden för ökad läsbarhet så att det blir tydligare vad som hänger ihop med vartannat.
