+++ weatherly/src/SmhiAPI.js
@@ -1,26 +1,34 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import axios from 'axios';
-import test from './static/day.svg'
+import Weather from './Weather';
 
 const baseURL = "https://opendata-download-metfcst.smhi.se/api/category/pmp3g/version/2/geotype/point/lon/16.158/lat/58.5812/data.json";
 export default function SmhiAPI(props) {
-    const [weather, setWeather] = React.useState([]);
-    const [geoLoc, setGeoloc] = React.useState([{"long": 12.233,
-"lat": 123.343}]);
+    const [weather, setWeather] = useState([]);
 
-    console.log(weather)
-    console.log(props.referenceTime)
+    //STÄDA BORT UTSKRIFT
+    console.log(weather);
 
-    React.useEffect(() => {
+    useEffect(() => {
+        const currentDate = new Date();
+        const dateTimeToString = currentDate.getFullYear() + "-" + ('0' + (currentDate.getMonth() + 1)).slice(-2) + "-" + currentDate.getDate() + "T" + currentDate.getHours() + ":00:00Z";
+        const weekday = currentDate.toLocaleDateString(undefined, { weekday: 'long' });
+        const capitalWeekday = weekday.slice(0, 1).toUpperCase() + weekday.slice(1, weekday.length)
+
         axios.get(baseURL).then((response) => {
             const SMHIData = [response.data]
+            //STÄDA BORT UTSKRIFT
+            console.log(SMHIData[0])
             const weatherData = SMHIData.map((tSerie) =>
                 tSerie.timeSeries.filter((time) =>
-                    time.validTime === props.referenceTime).map((params) =>
+                    time.validTime === dateTimeToString).map((params) =>
                         params.parameters.map((val) =>
                             val.values))
             );
+            //STÄDA BORT UTSKRIFT
+            console.log(weatherData + "DATA")
             const weatherParameters = {
+                "date": capitalWeekday,
                 "temperature": weatherData[0][0][10],
                 "windSpeed": weatherData[0][0][14],
                 "weatherCode": weatherData[0][0][18]
@@ -32,18 +40,24 @@ export default function SmhiAPI(props) {
 
     if (!weather) return null;
 
-    return (<>
-
-        <td>
-            <img src={test} alt="" />
-            {weather.weatherCode}
-        </td>
-        <td>
-            {weather.temperature}
-        </td>
-        <td>
-            {weather.windSpeed}m/s
-        </td>
-    </>
+    return (
+        <div>
+            <h2>Vädret just nu</h2>
+            <div>
+                <table className="table">
+                    <thead>
+                        <tr>
+                            <th scope="col">Dag</th>
+                            <th scope="col">Väder</th>
+                            <th scope="col">Temp</th>
+                            <th scope="col">Vind</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        <Weather weather={weather} />
+                    </tbody>
+                </table>
+            </div>
+        </div>
     )
 }
